Add files via upload
Done project for week 2- I have made an ingredient predictor model for an e-commerce site, which will predict the ingredients required for the toppings of the pizza by analyzing the photo provided by the user. 
I have used CNN (Convolution Neural Network) and used the NearestNeighbours algorithm to find the nearest match for the image provided by the user and the interface. Please unzip all the files in a single path.